SERVER-34006 build android embedded with api version 21

+++ buildscripts/setup-android-toolchain.sh
@@ -3,9 +3,9 @@
 set -o verbose
 set -o errexit
 
-_SystemImageArch=$1
+ToolchainArch=$1
 shift
-_ToolchainArch=$1
+API_VERSION=$1
 shift
 
 if [ -z "$PYTHON" ] ; then
@@ -13,22 +13,30 @@ if [ -z "$PYTHON" ] ; then
 fi
 
 SDK_ROOT=$PWD/android_sdk
-mkdir $SDK_ROOT
+if [ ! -e  $SDK_ROOT ]; then
+    mkdir $SDK_ROOT
+    (
+        cd $SDK_ROOT
+        SDK_PACKAGE=sdk-tools-linux-3859397.zip
+        curl -O https://dl.google.com/android/repository/$SDK_PACKAGE
+        unzip $SDK_PACKAGE
+        echo y | ./tools/bin/sdkmanager  \
+            "platforms;android-27"  \
+            "ndk-bundle"  \
+            "emulator"  \
+            "patcher;v4"  \
+            "platform-tools"  \
+            "build-tools;23.0.3"  \
+            "system-images;android-25;google_apis;armeabi-v7a"  \
+            "system-images;android-25;google_apis;arm64-v8a"  \
+            "system-images;android-21;google_apis;x86_64"
+    )
+fi
 
-TOOLCHAIN=$PWD/android_toolchain
-mkdir $TOOLCHAIN
+TOOLCHAIN=$PWD/android_toolchain-${ToolchainArch}-${API_VERSION}
+rm -rf $TOOLCHAIN
 
-API_VERSION=24
-
-(
-    cd $SDK_ROOT
-    SDK_PACKAGE=sdk-tools-linux-3859397.zip
-    curl -O https://dl.google.com/android/repository/$SDK_PACKAGE
-    unzip $SDK_PACKAGE
-    echo y | ./tools/bin/sdkmanager "platforms;android-24" "emulator" "ndk-bundle" "platform-tools" "build-tools;23.0.3" "system-images;android-24;google_apis;$_SystemImageArch"
-)
-
-$PYTHON $SDK_ROOT/ndk-bundle/build/tools/make_standalone_toolchain.py --arch $_ToolchainArch --api $API_VERSION  --stl=libc++ --force  --install-dir $TOOLCHAIN
+$PYTHON $SDK_ROOT/ndk-bundle/build/tools/make_standalone_toolchain.py --arch $ToolchainArch --api $API_VERSION --stl=libc++ --install-dir $TOOLCHAIN
 
 echo SDK_ROOT=${SDK_ROOT}
 echo TOOLCHAIN=${TOOLCHAIN}
